# contracts/openapi/platform-base.yaml
openapi: 3.1.0
info:
  title: Russian House Event Platform API
  version: 0.2.0
servers:
  - url: http://localhost:8080
paths:
...
  /v1/profile/me:
    get:
      summary: Read current profile
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Profile
    patch:
      summary: Update current profile
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                firstName:
                  type: string
                lastName:
                  type: string
                phone:
                  type: string
                birthDate:
                  type: string
                  format: date
      responses:
        '200':
          description: Updated profile
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT